.history/flaskapp/requirements_20240627122053.txt
Flask
marshmallow
unstructured
google-generativeai